Moved NWItem to core/item.py
